type change: size_t -> int due to < 0 comparison

+++ src/operations/transposition.cpp
@@ -79,7 +79,7 @@ void transposition::transpose ( string& pitch, float& alter, int& octave, int ta
 	int ialter = round(alter);
 	float diff = alter - ialter;
 	int pitch1 = notevisitor::step2i(pitch);
-	for (size_t i=0; i < fFifthCycle.size(); i++) {
+	for (int i=0; i < fFifthCycle.size(); i++) {
 		if ((fFifthCycle[i].second == ialter) && (fFifthCycle[i].first == pitch)) {
 			i += tableshift;
 			if (i > fFifthCycle.size()) i -= 12;
